Refactor #56 - Make `OnRoutine` and test package moved
