+++ app/src/main/java/com/nlab/reminder/ReminderApplication.kt
@@ -17,7 +17,7 @@
 package com.nlab.reminder
 
 import android.app.Application
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 import dagger.hilt.android.HiltAndroidApp
 
 /**
