+++ app/src/main/java/com/nlab/reminder/domain/common/effect/message/navigation/TagEndNavigationMessage.kt
@@ -17,7 +17,7 @@
 package com.nlab.reminder.domain.common.effect.message.navigation
 
 import com.nlab.reminder.core.effect.message.navigation.NavigationMessage
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 import com.nlab.reminder.domain.common.tag.Tag
 
 /**
