+++ app/src/main/java/com/nlab/reminder/domain/common/schedule/ScheduleUiState.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.domain.common.schedule
 
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 
 /**
  * @author Doohyun
