+++ app/src/main/java/com/nlab/reminder/domain/common/tag/Tag.kt
@@ -18,7 +18,7 @@ package com.nlab.reminder.domain.common.tag
 
 import android.os.Parcelable
 import androidx.annotation.Keep
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 import kotlinx.parcelize.Parcelize
 
 /**
