+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeState.kt
@@ -17,7 +17,7 @@
 package com.nlab.reminder.domain.feature.home
 
 import com.nlab.reminder.core.state.State
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 
 /**
  * @author Doohyun
