+++ app/src/main/java/com/nlab/reminder/domain/feature/home/tag/rename/HomeTagRenameState.kt
@@ -17,7 +17,7 @@
 package com.nlab.reminder.domain.feature.home.tag.rename
 
 import com.nlab.reminder.core.state.State
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 
 @Generated
 data class HomeTagRenameState(
