+++ app/src/main/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleReport.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.domain.feature.schedule.all
 
-import com.nlab.reminder.core.util.annotation.test.Generated
+import com.nlab.reminder.core.util.test.annotation.Generated
 import com.nlab.reminder.domain.common.schedule.ScheduleUiState
 
 /**
